+++ frontend/src/components/Tabs/index.tsx
@@ -1,23 +1,24 @@
-import { Tab } from './Tab';
-import { TabsProps } from './type';
-import { useLocation, useNavigate } from 'react-router-dom';
+import { TabProps, TabsProps } from './type';
+import useTabs from './useTabs';
 import styled, { css } from 'styled-components';
 
 import { FlexBox } from 'components';
 
+import COLOR from 'styles/color';
+
 export const Tabs = ({ tabList }: TabsProps) => {
-  const navigate = useNavigate();
-  const { pathname } = useLocation();
+  const { pathname, handleClickTab } = useTabs();
 
   return (
     <TabList as="ul" alignItems="center">
       {tabList.map(({ path, tabName }) => (
         <Tab
           key={path}
-          tabName={tabName}
           isSelected={path === pathname}
-          onClick={() => navigate(path)}
-        />
+          onClick={() => handleClickTab(path)}
+        >
+          {tabName}
+        </Tab>
       ))}
     </TabList>
   );
@@ -31,4 +32,21 @@ const TabList = styled(FlexBox)`
     margin-top: -7px;
     width: 100%;
   `}
+`;
+
+const Tab = styled.li<TabProps>`
+  ${({ isSelected, theme }) => `
+      cursor: pointer;
+      font-size: 17px;
+      padding: 1rem;
+      border-bottom: 2px solid ${COLOR.LIGHT_GRAY};
+      color: ${theme.onBackground};
+  
+      border-color: ${isSelected && theme.primary};
+      font-weight: ${isSelected ? 'bold' : 'normal'};
+  
+      &:hover {
+         border-color: ${theme.primary};
+      }
+    `}
 `;
